+++ CHANGELOG.md
@@ -5,8 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## 2.4.0
-- Added OAuthTokenRelatedClient
-- Added endpoint to delete access token
+- Added endpoint to revoke access token
 
 ## 2.3.2
 - Fixed deleteTransferConfiguration endpoint